# swagger.yaml
openapi: 3.0.0
paths:
...
  "/v1/users/{username}":
    get:
      operationId: findOneUserByUserame
      summary: Finds a user by :username
      parameters:
        - name: username
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbUser"
        "404":
          description: User not found
      tags:
        &a2
        - User service
...
  "/v1/users/{username}/highlights":
    get:
      operationId: findAllHighlightsByUsername
      summary: Listing all Highlights for a user and paginate them
      parameters:
        - name: username
          required: true
          in: path
          schema:
            type: string
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: integer
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 1000
            default: 10
            type: integer
        - name: orderDirection
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/OrderDirectionEnum"
        - name: range
          required: false
          in: query
          description: Range in days
          schema:
            default: 30
            type: integer
      responses:
        "200":
          description: Successfully received model list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageDto"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/DbUserHighlight"
        "404":
          description: Highlights not found
      tags: *a2
...
  /v1/user/highlights:
    post:
      operationId: addHighlightForUser
      summary: Adds a new highlight for the authenticated user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserHighlightDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OmitTypeClass"
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbUserHighlight"
        "400":
          description: Invalid request
        "404":
          description: Unable to add user highlight
      tags:
        &a3
        - User Highlights service
      security:
        - bearer: []
  "/v1/user/highlights/{id}":
    get:
      operationId: getUserHighlight
      summary: Retrieves a highlight
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: integer
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbUserHighlight"
        "400":
          description: Invalid request
        "404":
          description: Unable to get user highlight
      tags: *a3
    patch:
      operationId: updateHighlightForUser
      summary: Updates the highlight for the authenticated user
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserHighlightDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbUserHighlight"
        "400":
          description: Invalid request
        "404":
          description: Unable to update user highlight
      tags: *a3
      security:
        - bearer: []
    delete:
      operationId: deleteHighlightForUser
      summary: Deletes the highlight for the authenticated user
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: integer
      responses:
        "200":
          description: ""
        "400":
          description: Invalid request
        "404":
          description: Unable to delete user highlight
      tags: *a3
      security:
        - bearer: []
  "/v1/user/highlights/{id}/reactions":
    get:
      operationId: getAllHighlightUserReactions
      summary: Retrieves a highlight reactions for the authenticated user
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: integer
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbUserHighlightReactionResponse"
        "400":
          description: Invalid request
        "404":
          description: Unable to get user highlight reactions
      tags: *a3
      security:
        - bearer: []
  "/v1/user/highlights/{id}/reactions/{emojiId}":
    post:
      operationId: addHighlightReactionForUser
      summary: Adds a new highlight reaction for the authenticated user
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: integer
        - name: emojiId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbUserHighlightReaction"
        "201":
          description: ""
        "400":
          description: Invalid request
        "404":
          description: Highlight does not exist
        "409":
          description: Unable to add user highlight reaction
      tags: *a3
      security:
        - bearer: []
    delete:
      operationId: deleteHighlightReactionForUser
      summary: Deletes the highlight reaction for the authenticated user
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: integer
        - name: emojiId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
        "400":
          description: Invalid request
        "404":
          description: Unable to delete user highlight reaction
      tags: *a3
      security:
        - bearer: []
  /v1/user/highlights/following:
    get:
      operationId: getFollowingHighlights
      summary: Fetches highlights for users the authenticated user follows
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: integer
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 1000
            default: 10
            type: integer
        - name: orderDirection
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/OrderDirectionEnum"
        - name: range
          required: false
          in: query
          description: Range in days
          schema:
            default: 30
            type: integer
        - name: repo
          required: false
          in: query
          description: Highlight Repo Filter
          example: open-sauced/insights
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbUserHighlight"
      tags: *a3
      security:
        - bearer: []
  /v1/user/highlights/following/repos:
    get:
      operationId: getFollowingHighlightRepos
      summary: Fetches highlight repos for users the authenticated user follows
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: integer
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 1000
            default: 10
            type: integer
        - name: orderDirection
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/OrderDirectionEnum"
        - name: range
          required: false
          in: query
          description: Range in days
          schema:
            default: 30
            type: integer
        - name: repo
          required: false
          in: query
          description: Highlight Repo Filter
          example: open-sauced/insights
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbUserHighlightRepo"
        "400":
          description: Invalid request
      tags: *a3
      security:
        - bearer: []
...
  "/v1/repos/{id}":
    get:
      operationId: findOneById
      summary: Finds a repo by :id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: integer
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbRepo"
        "404":
          description: Repository not found
      tags:
        &a8
        - Repository service
...
  /v1/user/insights:
    get:
      operationId: findAllInsightsByUserId
      summary: Listing all insights for a user and paginate them
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: integer
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 1000
            default: 10
            type: integer
        - name: orderDirection
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/OrderDirectionEnum"
        - name: range
          required: false
          in: query
          description: Range in days
          schema:
            default: 30
            type: integer
      responses:
        "200":
          description: Successfully received model list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageDto"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/DbInsight"
        "404":
          description: Insights not found
      tags:
        &a22
        - Insights service
      security:
        - bearer: []
    post:
      operationId: addInsightForUser
      summary: Adds a new insight page for the authenticated user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInsightDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbInsight"
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbInsight"
        "400":
          description: Invalid request
        "404":
          description: Unable to add user insight
      tags: *a22
      security:
        - bearer: []
...
  /v1/highlights/repos/list:
    get:
      operationId: findAllHighlightRepos
      summary: Finds all highlight repos and paginates them
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: integer
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 1000
            default: 10
            type: integer
        - name: orderDirection
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/OrderDirectionEnum"
        - name: range
          required: false
          in: query
          description: Range in days
          schema:
            default: 30
            type: integer
      responses:
        "200":
          description: Successfully received model list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageDto"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/DbUserHighlightRepo"
      tags: *a24
  "/v1/highlights/{id}/reactions":
    get:
      operationId: getAllHighlightReactions
      summary: Retrieves total reactions for a highlight
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: integer
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbUserHighlightReactionResponse"
        "400":
          description: Invalid request
        "404":
          description: Unable to get user highlight reactions
      tags: *a24
...
info:
  title: "@open-sauced/api.opensauced.pizza"
  description: >-
    
    ## Swagger-UI API Documentation


    This REST API can be used to create, read, update or delete data from the Open Sauced community platform.

    The Swagger-UI provides useful information to get started and an overview of all available resources.

    Each API route is clickable and has their own detailed description on how to use it.

    The base URL for the API is [api.opensauced.pizza](https://api.opensauced.pizza).


    [comment]: # (TODO: add bearer auth information)


    ## Rate limiting


    Every IP address is allowed to perform 5000 requests per hour.

    This is measured by saving the date of the initial request and counting all requests in the next hour.

    When an IP address goes over the limit, HTTP status code 429 is returned.

    The returned HTTP headers of any API request show the current rate limit status:


    header | description

    --- | ---

    `X-RateLimit-Limit` | The maximum number of requests allowed per hour

    `X-RateLimit-Remaining` | The number of requests remaining in the current rate limit window

    `X-RateLimit-Reset` | The date and time at which the current rate limit window resets in [UTC epoch seconds](https://en.wikipedia.org/wiki/Unix_time)


    [comment]: # (TODO: add pagination information)


    ## Common response codes


...

    The table below shows most common response codes you can receive from our endpoints.


...

...

...

...

...

...

...

...

...

...


...
servers:
...
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    DbUser:
      type: object
      properties:
        id:
          type: integer
          description: User identifier
          example: 237133
        open_issues:
          type: integer
          description: Total number of open issues user has across public activity
          example: 498
        created_at:
          format: date-time
          type: string
          description: Timestamp representing user creation
          example: 2016-10-19 13:24:51.000000
        updated_at:
          format: date-time
          type: string
          description: Timestamp representing user last update
          example: 2022-08-28 22:04:29.000000
        first_opened_pr_at:
          format: date-time
          type: string
          description: Timestamp representing user first open PR
          example: 2022-08-28 22:04:29.000000
        first_pushed_commit_at:
          format: date-time
          type: string
          description: Timestamp representing user first commit push
          example: 2022-08-28 22:04:29.000000
        connected_at:
          format: date-time
          type: string
          description: Timestamp representing user logging in to open sauced for the first
            time
          example: 2022-08-28 22:04:29.000000
        node_id:
          type: string
          description: User GitHub node id
          example: MDQ6VXNlcjIzNzEzMw==
        avatar_url:
          type: string
          description: User GitHub avatar URL
          example: https://avatars.githubusercontent.com/u/237133?v=4
        gravatar_id:
          type: string
          description: User GitHub gravatar URL
          example: ""
        url:
          type: string
          description: User GitHub profile URL
          example: https://api.github.com/users/0-vortex
        login:
          type: string
          description: User unique login name
          example: 0-vortex
        email:
          type: string
          description: User email address
          example: hello@opensauced.pizza
        is_private:
          type: boolean
          description: Flag indicating whether user opted to have a private profile (beta
            feature
          example: false
          default: false
        is_open_sauced_member:
          type: boolean
          description: Flag indicating app.opensauced user status
          example: false
          default: false
        is_onboarded:
          type: boolean
          description: Flag indicating user's onboarding status
          example: false
          default: false
        is_waitlisted:
          type: boolean
          description: Flag indicating user's waitlist status
          example: false
          default: false
        role:
          type: integer
          description: Insights Role
          example: 10
          default: 10
        bio:
          type: string
          description: User bio information
          example: OpenSauced User
        blog:
          type: string
          description: GitHub blog information
          example: https://opensauced.pizza/blog
        name:
          type: string
          description: User name information
          example: MrPizza
        twitter_username:
          type: string
          description: User Twitter information
          example: saucedopen
        linkedin_url:
          type: string
          description: LinkedIn URL
          example: https://www.linkedin.com/in/brianldouglas
        github_sponsors_url:
          type: string
          description: GitHub Sponsors URL
          example: https://github.com/sponsors/open-sauced
        discord_url:
          type: string
          description: Discord URL
          example: https://discord.gg/opensauced
        company:
          type: string
          description: User company information
          example: OpenSauced
        location:
          type: string
          description: User location information
          example: San Francisco, CA
        display_local_time:
          type: boolean
          description: User display local time information
          example: false
        interests:
          type: string
          description: User topic interests
          example: javascript
        hireable:
          type: boolean
          description: GitHub user hireable status
          example: false
        public_repos:
          type: integer
          description: GitHub user public repos number
          example: 0
        public_gists:
          type: integer
          description: GitHub user public gists number
          example: 0
        followers:
          type: integer
          description: GitHub user public followers number
          example: 0
        following:
          type: integer
          description: GitHub user public following number
          example: 0
        type:
          type: string
          description: GitHub user type
          example: User
          default: User
        display_email:
          type: boolean
          description: User display public email
          example: false
        receive_collaboration:
          type: boolean
          description: User receives collaboration requests
          example: false
        timezone:
          type: string
          description: User timezone in UTC
          example: UTC-5
        languages:
          type: object
          description: GitHub top languages
          example: "{ TypeScript: 33128, HTML: 453, JavaScript: 90, CSS: 80 }"
          default: {}
        notification_count:
          type: integer
          description: User notification count
          example: 0
        highlights_count:
          type: integer
          description: User highlights count
          example: 0
        following_count:
          type: integer
          description: User following count
          example: 0
        followers_count:
          type: integer
          description: User followers count
          example: 0
        recent_pull_requests_count:
          type: integer
          description: Count of user pull requests within the last 30 days
          example: 0
        recent_pull_request_velocity_count:
          type: integer
          description: User average pull request velocity in days over the last 30 days
          example: 0
        is_maintainer:
          type: boolean
          description: Flag to indicate if user is a maintainer of any repos
          example: true
      required:
        - id
        - open_issues
        - node_id
        - avatar_url
        - login
        - email
        - is_private
        - is_open_sauced_member
        - is_onboarded
        - is_waitlisted
        - role
        - public_repos
        - public_gists
        - followers
        - following
        - type
        - languages
        - notification_count
        - highlights_count
        - following_count
        - followers_count
        - recent_pull_requests_count
        - recent_pull_request_velocity_count
        - is_maintainer
...
    PageDto:
      type: object
      properties:
        data:
          type: array
          items:
            type: array
        meta:
          $ref: "#/components/schemas/PageMetaDto"
      required:
        - data
        - meta
...
    OrderDirectionEnum:
      type: string
      enum:
        - ASC
        - DESC
    DbUserHighlight:
      type: object
      properties:
        id:
          type: integer
          description: User Highlight identifier
          example: 237133
        user_id:
          type: integer
          description: User identifier
          example: 237133
        url:
          type: string
          description: Highlight Pull Request URL
          example: github.com/open-sauced/insights/pull/1
        title:
          type: string
          description: Highlight Title
          example: My First PR!
        highlight:
          type: string
          description: Highlight Text
          example: |-2
            
                I made my first open source pull request!

                github.com/open-sauced/insights/pull/1
        type:
          type: string
          description: Highlight type
          example: pull_request
        pinned:
          type: boolean
          description: Whether the highlight is pinned to the top
          example: false
        featured:
          type: boolean
          description: Whether the highlight is featured or not
          example: false
        created_at:
          format: date-time
          type: string
          description: Timestamp representing highlight creation date
          example: 2023-01-19 13:24:51.000000
        updated_at:
          format: date-time
          type: string
          description: Timestamp representing highlight updated date
          example: 2023-01-19 13:24:51.000000
        deleted_at:
          format: date-time
          type: string
          description: Timestamp representing highlight deletion date
          example: 2023-01-19 13:24:51.000000
        shipped_at:
          format: date-time
          type: string
          description: Timestamp representing highlight shipped date
          example: 2023-01-19 13:24:51.000000
        full_name:
          type: string
          description: Highlight Repo Full Name
          example: open-sauced/insights
        name:
          type: string
          description: Highlight User Full Name
          example: Brian Douglas
        login:
          type: string
          description: Highlight User Login
          example: bdougie
        tagged_repos:
          description: An array of full-names of tagged repositories
          example:
            - open-sauced/insights
            - open-sauced/ai
          type: array
          items:
            type: string
      required:
        - id
        - user_id
        - highlight
        - type
        - tagged_repos
...
    CreateUserHighlightDto:
      type: object
      properties:
        url:
          type: string
          description: Highlight PR URL
          example: github.com/open-sauced/insights/pull/1
        title:
          type: string
          description: Highlight Title
          example: My first PR!
        highlight:
          type: string
          description: Highlight Text
          example: My first PR to Open Sauced!
        shipped_at:
          type: string
          description: Shipped Date
          example: 2023-01-19 13:24:51.000000
        type:
          type: string
          default: pull_request
          description: Highlight type
          example: pull_request
        taggedRepos:
          description: An array of full-names of tagged repositories
          example:
            - open-sauced/insights
            - open-sauced/ai
          type: array
          items:
            type: string
      required:
        - url
        - highlight
        - type
        - taggedRepos
    OmitTypeClass:
      type: object
      properties:
        id:
          type: integer
          description: Notification identifier
          example: 237133
        user_id:
          type: integer
          description: User ID
          example: 498
        from_user_id:
          type: integer
          description: From User ID
          example: 43311
        type:
          type: string
          description: User notification type
          example: welcome
        message:
          type: string
          description: User notification message
          example: bdougie followed you
        notified_at:
          format: date-time
          type: string
          description: Timestamp representing db-user-notification creation
          example: 2022-10-19 13:24:51.000000
        meta_id:
          type: string
          description: Notification Source ID (highlight, user, invite)
          example: "133"
      required:
        - id
        - user_id
        - type
...
    DbUserHighlightRepo:
      type: object
      properties:
        full_name:
          type: string
          description: Highlight Repo Full Name
          example: open-sauced/insights
...